requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.44
alembic==1.16.5
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.12.3
pydantic[email]==2.12.3
pdf2image==1.17.0
Pillow==11.3.0
python-multipart==0.0.20
python-dotenv==1.1.1
email-validator==2.3.0
bcrypt==4.0.1
python-pptx==0.6.23
